+++ lib/src/cli/commands/reply_command.dart
@@ -1,15 +1,18 @@
 import 'dart:io';
 
 import 'package:dart_slack/src/cli/commands/authenticated_command.dart';
-import 'package:dart_slack/src/cli/message_text.dart';
+import 'package:dart_slack/src/cli/message_text_input.dart';
 import 'package:dart_slack/src/slack.dart';
 import 'package:mason_logger/mason_logger.dart';
 
-/// `dart_slack reply --channel <id> --thread <ts> --text "message"`
+/// `dart_slack reply --channel <id> --thread <ts> --text-file <path>`
 /// `[--file <path>]`
 ///
+/// The reply text comes from exactly one of `--text`, `--text-file` or
+/// `--text-stdin`. `--file` attaches a file and is a different option.
+///
 /// Replies to a message thread in a Slack channel.
-class ReplyCommand extends AuthenticatedCommand {
+class ReplyCommand extends AuthenticatedCommand with MessageTextCommand {
   /// Creates a [ReplyCommand].
   ReplyCommand({
     required super.logger,
@@ -29,17 +32,8 @@ class ReplyCommand extends AuthenticatedCommand {
         help: 'The thread timestamp (thread_ts) to reply to.',
         mandatory: true,
       )
-      ..addOption(
-        'text',
-        abbr: 't',
-        help: 'The reply text.',
-        mandatory: true,
-      )
-      ..addOption(
-        'file',
-        abbr: 'f',
-        help: 'Path to a local file to attach.',
-      );
+      ..addOption('file', abbr: 'f', help: 'Path to a local file to attach.');
+    addMessageTextOptions(argParser, help: 'The reply text.');
   }
 
   @override
@@ -52,7 +46,6 @@ class ReplyCommand extends AuthenticatedCommand {
   Future<int> runAuthenticated(Slack slack) async {
     final channel = argResults!['channel'] as String;
     final thread = argResults!['thread'] as String;
-    final text = normalizeMessageText(argResults!['text'] as String);
     final filePath = argResults!['file'] as String?;
 
     if (filePath != null) {
@@ -64,7 +57,7 @@ class ReplyCommand extends AuthenticatedCommand {
         channel: channel,
         path: filePath,
         threadTs: thread,
-        comment: text,
+        comment: messageText,
       );
       logger.success('File "$filename" sent to thread $thread in $channel.');
       return ExitCode.success.code;
@@ -72,7 +65,7 @@ class ReplyCommand extends AuthenticatedCommand {
 
     final message = await slack.postMessage(
       channel: channel,
-      text: text,
+      text: messageText,
       threadTs: thread,
     );
     logger.success(